+++ helpers/env-variables.ts
@@ -5,6 +5,7 @@ import {
   ModelConfig,
   TemplateDataSource,
   TemplateFramework,
+  TemplateType,
   TemplateVectorDB,
 } from "./types";
 
@@ -264,6 +265,15 @@ const getModelEnvs = (modelConfig: ModelConfig): EnvVar[] => {
           },
         ]
       : []),
+    ...(modelConfig.provider === "mistral"
+      ? [
+          {
+            name: "MISTRAL_API_KEY",
+            description: "The Mistral API key to use.",
+            value: modelConfig.apiKey,
+          },
+        ]
+      : []),
     ...(modelConfig.provider === "t-systems"
       ? [
           {
@@ -378,6 +388,36 @@ const getSystemPromptEnv = (tools?: Tool[]): EnvVar => {
   };
 };
 
+const getTemplateEnvs = (template?: TemplateType): EnvVar[] => {
+  if (template === "multiagent") {
+    return [
+      {
+        name: "MESSAGE_QUEUE_PORT",
+      },
+      {
+        name: "CONTROL_PLANE_PORT",
+      },
+      {
+        name: "HUMAN_CONSUMER_PORT",
+      },
+      {
+        name: "AGENT_QUERY_ENGINE_PORT",
+        value: "8003",
+      },
+      {
+        name: "AGENT_QUERY_ENGINE_DESCRIPTION",
+        value: "Query information from the provided data",
+      },
+      {
+        name: "AGENT_DUMMY_PORT",
+        value: "8004",
+      },
+    ];
+  } else {
+    return [];
+  }
+};
+
 export const createBackendEnvFile = async (
   root: string,
   opts: {
@@ -386,6 +426,7 @@ export const createBackendEnvFile = async (
     modelConfig: ModelConfig;
     framework: TemplateFramework;
     dataSources?: TemplateDataSource[];
+    template?: TemplateType;
     port?: number;
     tools?: Tool[];
   },
@@ -406,6 +447,8 @@ export const createBackendEnvFile = async (
     ...getVectorDBEnvs(opts.vectorDb, opts.framework),
     ...getFrameworkEnvs(opts.framework, opts.port),
     ...getToolEnvs(opts.tools),
+    // Add template environment variables
+    ...getTemplateEnvs(opts.template),
     getSystemPromptEnv(opts.tools),
   ];
   // Render and write env file